+++ Sources/OnboardingKit/Views/OnboardingUspListItemStyle.swift
@@ -1,0 +1,43 @@
+//
+//  OnboardingUspListItemStyle.swift
+//  OnboardingKit
+//
+//  Created by Daniel Saidi on 2025-11-21.
+//  Copyright © 2025 Daniel Saidi. All rights reserved.
+//
+
+import SwiftUI
+
+/// This style can be used with ``OnboardingIntroScreen``.
+///
+/// This style can be applied with``SwiftUICore/View/OnboardingUspListItemStyle(_:)`.
+public struct OnboardingUspListItemStyle {
+
+    public init(
+        iconSize: Double = 45
+    ) {
+        self.iconSize = iconSize
+    }
+
+    public let iconSize: Double
+}
+
+public extension OnboardingUspListItemStyle {
+
+    /// The standard intro screen style.
+    static var standard: Self { .init() }
+}
+
+public extension EnvironmentValues {
+
+    @Entry var onboardingUspListItemStyle = OnboardingUspListItemStyle()
+}
+
+public extension View {
+
+    func onboardingUspListItemStyle(
+        _ style: OnboardingUspListItemStyle
+    ) -> some View {
+        self.environment(\.onboardingUspListItemStyle, style)
+    }
+}